# [8_10] list: perf tuning on list_contains slightly