--- a/bench/Kernel/Containers/list_bench.cpp
+++ b/bench/Kernel/Containers/list_bench.cpp
@@ -25,6 +25,7 @@ main () {
   list<long> l16  = gen (16);
   list<long> l32  = gen (32);
   list<long> l32_2= gen (32);
+  list<long> l64  = gen (64);
 
   bench.run ("last_item 1", [&] { last_item (l1); });
   bench.run ("last_item 4", [&] { last_item (l4); });
@@ -32,11 +33,12 @@ main () {
 
   bench.run ("N 32", [&] { N (l32); });
 
-  bench.minEpochIterations (40000);
-
+  bench.minEpochIterations (100000);
   bench.run ("contains 1", [&] {contains (l32, 1L); });
   bench.run ("contains 16", [&] {contains (l32, 16L); });
   bench.run ("contains 32", [&] {contains (l32, 32L); });
+  bench.run ("contains 64", [&] {contains (l32, 64L); });
 
+  bench.minEpochIterations (40000);
   bench.run ("l1 == l2 32", [&] { l32 == l32_2; });
 }